refactor: split RunnerPoller.swift into four focused files
- RunnerPoller.swift — declaration, stored properties, init, deinit, setDisplayState
- RunnerPoller+PollLoop.swift — startObservingScopes, start, updateAdaptiveCounters,
  hasActiveWork, nextPollInterval
- RunnerPoller+FetchInternal.swift — fetch, fetchInternal, deriveExtraOrgScopes
- RunnerPoller+FetchHelpers.swift — fetchAllJobs, fetchActionGroups

All code and comments carried over verbatim. SwiftLint file_length now satisfied
across all four files. No logic changes.
