+++ Sources/RunBotCore/Runner/Polling/RunnerPoller.swift
@@ -1,5 +1,15 @@
 // RunnerPoller.swift
 // RunBotCore
+//
+// Actor declaration, stored properties, init/deinit, and the setDisplayState
+// write-through helper. All method implementations live in extension files:
+//
+//   RunnerPoller+PollLoop.swift      — observation loop, start(), interval math
+//   RunnerPoller+FetchInternal.swift — fetch(), fetchInternal(), deriveExtraOrgScopes()
+//   RunnerPoller+FetchHelpers.swift  — fetchAllJobs(), fetchActionGroups()
+//   RunnerPoller+ApplyResult.swift   — applyFetchResult(), applyError()
+//   RunnerPoller+PollBridge.swift    — buildJobState(), buildGroupState()
+
 import Foundation
 import GitHubClient
 import os
@@ -204,313 +214,6 @@ public actor RunnerPoller {
     pollLoop.cancelAll()
     let prefetchQueue = zipPrefetchQueue
     Task { await prefetchQueue.cancelAll() }
-  }
-
-  // MARK: - Observation loops
-
-  /// Starts (or restarts) the `activeScopes` observation loop.
-  ///
-  /// **Self-cancellation avoidance**
-  /// The new `Task` is created first, then handed to `setScopeObservationTask` so
-  /// the setter cancels the *previous* task rather than the one currently executing.
-  private func startObservingScopes() {
-    let injectedStore = scopeStore
-    let newTask = Task { [weak self] in
-      let (stream, continuation) = AsyncStream<[String]>.makeStream()
-      let observer: ObservationRelay<[String]> = await MainActor.run {
-        let relay = ObservationRelay<[String]>(continuation: continuation) {
-          injectedStore.activeScopes
-        }
-        relay.start()
-        return relay
-      }
-      for await _ in stream {
-        guard !Task.isCancelled else { break }
-        log("RunnerPoller › ScopeStore.activeScopes changed — restarting fetch", category: .runner)
-        await self?.startObservingScopes()
-        guard !Task.isCancelled else { break }
-        await self?.start()
-        break
-      }
-      // LOAD-BEARING: keeps the ObservationRelay alive until the for-await loop above
-      // exits. Without this, ARC may drop `observer` immediately after the `let`
-      // binding above goes out of scope (the Task captures `self` weakly and the relay
-      // is not otherwise retained), silently stopping scope-change detection.
-      withExtendedLifetime(observer) {}
-    }
-    pollLoop.setScopeObservationTask(newTask)
-  }
-
-  // MARK: - Poll loop
-
-  /// Starts (or restarts) the structured async poll loop.
-  public func start() async {
-    // Reset adaptive-interval counters so every restart begins from a clean state.
-    // `rateLimitRemaining` is intentionally NOT reset: a stale-but-real remaining
-    // value from before the restart is more useful than the Int.max sentinel.
-    consecutiveIdleTicks = 0
-    lastBusyRunnerCount = 0
-    // Clear prevLiveJobs and completedCache so stale entries from the previous scope
-    // do not trigger duplicate notifications after a scope-change restart.
-    // Jobs completing in the narrow window between this clear and the next cycle's
-    // API response cannot be notified — accepted pre-existing behaviour.
-    prevLiveJobs = [:]
-    completedCache = [:]
-    let scopes = await MainActor.run { scopeStore.activeScopes }
-    log("RunnerPoller › start — activeScopes=\(scopes)", category: .runner)
-    if scopes.isEmpty {
-      log(
-        "RunnerPoller › ⚠️ start called but activeScopes is EMPTY — actions will not load",
-        category: .runner)
-    }
-    let localCount = await MainActor.run { localRunners().count }
-    log(
-      "RunnerPoller › start — localRunners.count=\(localCount) at start() time", category: .runner)
-    if localCount == 0 {
-      log(
-        "RunnerPoller › ⚠️ start — localRunners=0; installPathMap will be empty on first fetch.",
-        category: .runner)
-    }
-    log(
-      "RunnerPoller › start — previous pollTask cancelled, launching new poll task",
-      category: .runner)
-    // [weak self] breaks the strong actor → pollLoop → Task → actor retain cycle.
-    pollLoop.setPollTask(
-      Task { [weak self] in
-        await self?.fetch()
-        while let self, !Task.isCancelled {
-          let interval = await self.nextPollInterval()
-          log("RunnerPoller › poll loop — next fetch in \(Int(interval))s", category: .runner)
-          do {
-            try await Task.sleep(for: .seconds(interval))
-          } catch is CancellationError {
-            log("RunnerPoller › poll loop — CancellationError, exiting cleanly", category: .runner)
-            break
-          } catch {
-            log("RunnerPoller › poll loop — unexpected error \(error), exiting", category: .runner)
-            break
-          }
-          guard !Task.isCancelled else {
-            log("RunnerPoller › poll loop — cancelled after sleep, exiting", category: .runner)
-            break
-          }
-          await self.fetch()
-        }
-      })
-  }
-
-  /// Updates the two adaptive-interval counters after a successful fetch cycle.
-  ///
-  /// ⚠️ **Ordering constraint — call site is `RunnerPoller+ApplyResult.swift` only**:
-  /// Must be called *after* `setDisplayState` so that `hasActiveWork()` evaluates
-  /// the freshly-written `self.jobs` / `self.actions`.
-  ///
-  /// - Returns: The updated `consecutiveIdleTicks` value for log interpolation.
-  @discardableResult
-  func updateAdaptiveCounters(hasActiveWork: Bool, busyRunnerCount: Int) -> Int {
-    if hasActiveWork {
-      consecutiveIdleTicks = 0
-    } else {
-      consecutiveIdleTicks += 1
-    }
-    lastBusyRunnerCount = busyRunnerCount
-    return consecutiveIdleTicks
-  }
-
-  /// Returns `true` when at least one job or action group is currently active
-  /// (in-progress or queued).
-  func hasActiveWork() -> Bool {
-    let hasActiveJobs = jobs.contains { $0.jobStatus == .inProgress || $0.jobStatus == .queued }
-    let hasActiveActions = actions.contains {
-      $0.groupStatus == .inProgress || $0.groupStatus == .queued
-    }
-    return hasActiveJobs || hasActiveActions
-  }
-
-  /// Computes the delay before the next poll by delegating to `PollIntervalStrategy`.
-  private func nextPollInterval() -> TimeInterval {
-    let hasActive = hasActiveWork()
-    let interval = PollIntervalStrategy.next(
-      hasActiveWork: hasActive,
-      consecutiveIdleTicks: consecutiveIdleTicks,
-      busyRunnerCount: lastBusyRunnerCount,
-      isRateLimited: isRateLimited,
-      rateLimitResetDate: rateLimitResetDate,
-      rateLimitRemaining: rateLimitRemaining
-    )
-    log(
-      "RunnerPoller › nextPollInterval — \(Int(interval))s hasActive=\(hasActive) idleTick=\(consecutiveIdleTicks) busyRunners=\(lastBusyRunnerCount) rateLimited=\(isRateLimited) rateLimitRemaining=\(rateLimitRemaining)",
-      category: .runner)
-    return interval
-  }
-
-  // MARK: - Fetch
-
-  /// Performs one full poll cycle.
-  func fetch() async {
-    do {
-      try await fetchInternal()
-    } catch {
-      log("RunnerPoller › fetch — ⚠️ unhandled error: \(error)", category: .runner)
-      await applyError(FetchError(error))
-    }
-  }
-
-  /// Inner throwing fetch body.
-  private func fetchInternal() async throws {
-    await clearGhRateLimit()
-    let scopesSnapshot = await MainActor.run { scopeStore.activeScopes }
-    log("RunnerPoller › fetch ENTER — activeScopesSnapshot=\(scopesSnapshot)", category: .runner)
-    if scopesSnapshot.isEmpty {
-      log("RunnerPoller › ⚠️ fetch — activeScopes snapshot is EMPTY", category: .runner)
-    }
-    let snapPrev = prevLiveJobs
-    let snapCache = completedCache
-    let snapPrevGroups = prevLiveGroups
-    let snapGroupCache = actionGroupCache
-    let localRunnersSnapshot = await MainActor.run { localRunners() }
-    log(
-      "RunnerPoller › fetch — localRunners.count=\(localRunnersSnapshot.count) (total; used for enrichment)",
-      category: .runner)
-    if localRunnersSnapshot.isEmpty {
-      log(
-        "RunnerPoller › ⚠️ fetch — localRunners is EMPTY; installPathMap will be empty",
-        category: .runner)
-    } else {
-      #if DEBUG
-        log(
-          "RunnerPoller › fetch — localRunners=\(localRunnersSnapshot.map { "\($0.runnerName)(agentId=\(String(describing: $0.agentId)) apiId=\(String(describing: $0.apiId)))" })",
-          category: .runner)
-      #endif
-    }
-    // (#2436) Filter to only runners with a resolved apiId — runners without one
-    // cannot contribute useful entries to InstallPathMap and silently pollute it.
-    // Full `localRunnersSnapshot` is still passed to `fetchAndEnrichRunners` below
-    // so newly-registered runners can have their apiId resolved on the next cycle.
-    let enabledLocalRunners = localRunnersSnapshot.filter { $0.apiId != nil }
-    log(
-      "RunnerPoller › fetch — enabledLocalRunners.count=\(enabledLocalRunners.count) (apiId resolved; used for installPathMap) of \(localRunnersSnapshot.count) total",
-      category: .runner)
-    if enabledLocalRunners.isEmpty && !localRunnersSnapshot.isEmpty {
-      log(
-        "RunnerPoller › ⚠️ fetch — no local runners with resolved apiId; installPathMap empty this cycle.",
-        category: .runner)
-    }
-    // Derive extra org scopes before buildInstallPathMap so byFullKey covers
-    // inferred org scopes as well as user-configured ones.
-    let extraOrgScopes = deriveExtraOrgScopes(
-      from: enabledLocalRunners,
-      configuredScopes: scopesSnapshot
-    )
-    log(
-      "RunnerPoller › fetch — extraOrgScopes=\(extraOrgScopes) (\(extraOrgScopes.count) inferred from local runner gitHubUrl)",
-      category: .runner)
-    let allScopes = scopesSnapshot + extraOrgScopes
-    let installPathMap = buildInstallPathMap(
-      scopes: allScopes,
-      localRunners: enabledLocalRunners
-    )
-    let enrichedRunners = await fetchAndEnrichRunners(
-      scopes: scopesSnapshot,
-      extraOrgScopes: extraOrgScopes,
-      localRunners: localRunnersSnapshot,
-      installPathMap: installPathMap
-    )
-    let jobResult = await buildJobState(
-      snapPrev: snapPrev,
-      snapCache: snapCache,
-      scopes: scopesSnapshot
-    )
-    let groupResult = await buildGroupState(
-      snapPrevGroups: snapPrevGroups,
-      snapGroupCache: snapGroupCache,
-      jobCache: jobResult.newCache,
-      scopes: scopesSnapshot
-    )
-    let didPickUp = await applyFetchResult(
-      enrichedRunners: enrichedRunners,
-      jobResult: jobResult,
-      groupResult: groupResult
-    )
-    // (#2327) Restart poll loop on runner pickup so the next fetch uses a fresh
-    // activeScopes snapshot. start() is called after applyFetchResult returns
-    // so notification dispatch completes before the cache is cleared.
-    if didPickUp { await start() }
-  }
-
-  /// Derives extra org scopes from local runner `gitHubUrl` values that are not
-  /// already present in the user-configured scope list.
-  ///
-  /// Only org-scoped URLs (single path component, no "/" in the derived scope)
-  /// are returned. Repo-scoped URLs are filtered out by the `!contains("/")` guard.
-  /// Duplicates and scopes already in `configuredScopes` are suppressed.
-  func deriveExtraOrgScopes(
-    from localRunners: [RunnerModel],
-    configuredScopes: [String]
-  ) -> [String] {
-    let configuredScopeSet = Set(configuredScopes)
-    var extraSet = Set<String>()
-    var extra: [String] = []
-    for localRunner in localRunners {
-      guard let url = localRunner.gitHubUrl,
-        let derivedScope = scopeFromUrl(url),
-        !derivedScope.contains("/"),
-        !configuredScopeSet.contains(derivedScope),
-        extraSet.insert(derivedScope).inserted
-      else { continue }
-      extra.append(derivedScope)
-      log(
-        "RunnerPoller › deriveExtraOrgScopes — derived '\(derivedScope)' from '\(localRunner.runnerName)'",
-        category: .runner)
-    }
-    return extra
-  }
-
-  /// Fetches all active jobs across all scopes concurrently, injecting the source scope
-  /// into each job.
-  ///
-  /// - Parameter scopes: The scope snapshot captured by `fetchInternal` — passed in
-  ///   directly to avoid re-reading `scopeStore.activeScopes` and creating a TOCTOU window.
-  ///
-  /// `internal` — required for cross-file extension access from `RunnerPoller+PollBridge.swift`.
-  func fetchAllJobs(scopes: [String]) async -> [ActiveJob] {
-    guard !scopes.isEmpty else { return [] }
-    var allJobs: [ActiveJob] = []
-    await withTaskGroup(of: [ActiveJob].self) { group in
-      for scope in scopes {
-        group.addTask {
-          await fetchActiveJobs(for: scope)
-            .map { $0.copying(scope: scope) }
-        }
-      }
-      for await jobs in group { allJobs.append(contentsOf: jobs) }
-    }
-    log(
-      "RunnerPoller › fetchAllJobs — fetched \(allJobs.count) job(s) across \(scopes.count) scope(s)",
-      category: .runner)
-    return allJobs
-  }
-
-  /// Fetches workflow action groups for the given scopes concurrently.
-  ///
-  /// - Parameter scopes: The scope snapshot captured by `fetchInternal` — passed in
-  ///   directly to avoid re-reading `scopeStore.activeScopes` and creating a TOCTOU window.
-  ///
-  /// `internal` — required for cross-file extension access from `RunnerPoller+PollBridge.swift`.
-  func fetchActionGroups(scopes: [String], shaKeyedCache: [String: WorkflowActionGroup]) async
-    -> [WorkflowActionGroup] {
-    guard !scopes.isEmpty else { return [] }
-    var allGroups: [WorkflowActionGroup] = []
-    await withTaskGroup(of: [WorkflowActionGroup].self) { group in
-      for scope in scopes {
-        group.addTask { await self.actionGroupFetcher.fetch(for: scope, cache: shaKeyedCache) }
-      }
-      for await groups in group { allGroups.append(contentsOf: groups) }
-    }
-    log(
-      "RunnerPoller › fetchActionGroups — fetched \(allGroups.count) group(s) across \(scopes.count) scope(s)",
-      category: .runner)
-    return allGroups
   }
 
   // MARK: - Private(set) write-through
